extracting players logic
