+++ tic-tac-toe game/tic_tac_toe.py
@@ -1,3 +1,5 @@
+from players import *
+
 def display_board(board):
     spacing = "-"*11
     print(f"+{spacing}+")
@@ -7,37 +9,6 @@ def display_board(board):
     print(f"|{spacing}|")
     print(f"| {board[6]} | {board[7]} | {board[8]} |")
     print(f"+{spacing}+")
-
-
-def get_player_name(player_no):
-    return input(f"Player {player_no}, please enter your name: ")
-
-
-def get_player_sign():
-    char = None
-    accepted_chars = ['X', 'O']
-    while char not in accepted_chars:
-        char = input("Which sign do you prefer? 'X' or 'O'? ").upper()
-        if char not in accepted_chars:
-            print("Invalid choice. Please choose 'X' or 'O'.")
-    return char
-
-
-def create_players():
-    player1 = {
-        'name': get_player_name(1),
-        'sign': get_player_sign()
-    }
-
-    player2 = {
-        'name': get_player_name(2),
-        'sign': 'O' if player1['sign'] == 'X' else 'X'
-    }
-    return {'player1': player1, 'player2': player2}
-
-
-def switch_player(current_player):
-    return players['player2'] if current_player == players['player1'] else players['player1']
 
 
 def get_available_fields(board):
@@ -83,6 +54,6 @@ for i in range(9):
         print(f"{player['name']} has won. Congratulations!")
         break
     else:
-        player = switch_player(player)
+        player = switch_player(player, players)
 else:
     print("It's a draw!")
